Add NeoFormat to Neovim config.
Adds the Neoformat plugin to the set of neovim tools.  Binds ,nf in both
normal and visual mode to trigger Neoformat, which itself can be
configured to leverage a massive variety of formatting tools.

This diff also attempts to setup Yapf to format python code.  It will
attempt to load a yapf style config from ~/.style.yapf.